# public/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ESG Hub API",
    "version": "1.0.0",
    "description": "Public API for the Ascent Partners Foundation ESG Knowledge Hub. Access 354 ESG articles and 244 curated external resources covering Environmental, Social, and Governance topics.",
    "contact": {
      "name": "ESG Hub Developers",
      "url": "https://esg-hub.ascent.partners/developers"
    },
    "license": {
      "name": "CC BY-SA 4.0",
      "url": "https://creativecommons.org/licenses/by-sa/4.0/"
    }
  },
  "servers": [
    {
      "url": "https://esg-hub.ascent.partners",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/search": {
      "get": {
        "summary": "Keyword or hybrid search",
        "description": "Full-text keyword search (BM25) or hybrid search (BM25 + HNSW with ESG re-ranking).",
        "operationId": "search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "maxLength": 500 },
            "description": "Search query"
          },
          {
            "name": "mode",
            "in": "query",
            "schema": { "type": "string", "enum": ["keyword", "hybrid"], "default": "keyword" }
          },
          {
            "name": "source",
            "in": "query",
            "schema": { "type": "string", "enum": ["all", "pages", "external"], "default": "all" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 50 }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results with ranking",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": { "type": "string" },
                    "mode": { "type": "string" },
                    "results": { "type": "array", "items": { "type": "object" } },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Semantic vector search",
        "description": "Search using a pre-computed embedding vector (384-dim, BGE-small).",
        "operationId": "semanticSearch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["embedding"],
                "properties": {
                  "embedding": {
                    "type": "array",
                    "items": { "type": "number" },
                    "minItems": 384,
                    "maxItems": 384
                  },
                  "k": { "type": "integer", "default": 10, "minimum": 1, "maximum": 50 },
                  "source": { "type": "string", "enum": ["all", "pages", "external"], "default": "all" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Semantic search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mode": { "type": "string" },
                    "model": { "type": "string" },
                    "dimensions": { "type": "integer" },
                    "data": { "type": "array", "items": { "type": "object" } },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "ESG Hub Write Token (required for POST/PATCH endpoints)"
      }
    }
  },
...
}